tugas crud & access

+++ src/tugas/tugas.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { User } from '@prisma/client';
+import { Role, User } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
 import * as path from 'path';
 import * as fs from 'fs';
@@ -8,36 +8,31 @@ import * as fs from 'fs';
 export class TugasService {
     constructor(private prisma: PrismaService) { }
 
-    async getTugas(user: User, id: number, req: any) {
+    async getTugas(user: User, pertemuanId: number, req: any) {
         try {
-            const tugas = await this.prisma.tugas.findUnique({
-                where: {
-                    id: id,
-                },
-            })
+            const tugas = await this.prisma.tugas.findMany({ where: { pertemuanId: pertemuanId } })
+            if (tugas.length === 0) throw new NotFoundException('Tugas tidak ditemukan');
+            const tugasWithFile = tugas.map((tugas) => {
+                if (tugas.file != null) tugas.file = `${req.protocol}://${req.headers.host}/tugas/${tugas.file}`;
+                else tugas.file = null;
+                return tugas;
+            });
+            return tugasWithFile;
+        } catch (error) { throw error }
+    }
 
+    async getTugasById(user: User, id: number, req: any) {
+        try {
+            const tugas = await this.prisma.tugas.findUnique({ where: { id: id } })
             if (!tugas) throw new NotFoundException('Tugas tidak ditemukan');
-
-            if (tugas.file != null) {
-                tugas.file = `http://${req.headers.host}/public/tugas/${tugas.file}`;
-            }
-
+            if (tugas.file != null) tugas.file = `${req.protocol}://${req.headers.host}/tugas/${tugas.file}`;
+            else tugas.file = null;
             return tugas;
-        } catch (error) {
-            throw error;
-        }
+        } catch (error) { throw error }
     }
 
     async addTugas(data: any, id: number) {
         try {
-            const tugasExist = await this.prisma.tugas.findFirst({
-                where: {
-                    pertemuanId: id,
-                }
-            });
-
-            if (tugasExist) throw new NotFoundException('Tugas sudah ada');
-
             const tugas = await this.prisma.tugas.create({
                 data: {
                     file: data.file,
@@ -47,32 +42,23 @@ export class TugasService {
                     pertemuanId: id,
                 }
             });
-
             if (tugas.file) {
                 const filePath = path.join(__dirname, '..', '..', 'public', 'tugas', tugas.file);
             }
-
             return tugas;
-        } catch (error) {
-            throw error;
-        }
+        } catch (error) { throw error }
     }
 
-    async updateTugas(data: any, id: number) {
+    async updateTugas(user: User, data: any, id: number) {
         try {
-            const tugas = await this.prisma.tugas.findUnique({
-                where: {
-                    id: id,
-                },
-            });
-
+            if (user.role == Role.MAHASISWA) throw new NotFoundException('Anda tidak memiliki akses');
+            const tugas = await this.prisma.tugas.findUnique({ where: { id: id } });
             if (tugas.file) {
                 const filePath = path.join(__dirname, '..', '..', 'public', 'tugas', tugas.file);
                 fs.unlink(filePath, (err) => {
                     if (err) throw err;
                 });
             }
-
             const tugasUpdate = await this.prisma.tugas.update({
                 data: {
                     file: data.file,
@@ -80,14 +66,25 @@ export class TugasService {
                     deskripsi: data.deskripsi,
                     tanggal: data.tanggal,
                 },
-                where: {
-                    id: id,
-                }
+                where: { id: id }
             });
-
             return tugasUpdate;
-        } catch (error) {
-            throw error;
-        }
+        } catch (error) { throw error }
+    }
+
+    async removeTugas(user: User, id: number) {
+        try {
+            if (user.role == Role.MAHASISWA) throw new NotFoundException('Anda tidak memiliki akses');
+            const tugas = await this.prisma.tugas.findUnique({ where: { id: id }, });
+            if (!tugas) throw new NotFoundException('Tugas tidak ditemukan');
+            if (tugas.file) {
+                const filePath = path.join(__dirname, '..', '..', 'public', 'tugas', tugas.file);
+                fs.unlink(filePath, (err) => {
+                    if (err) throw err;
+                });
+            }
+            const tugasDelete = await this.prisma.tugas.delete({ where: { id: id } });
+            return tugasDelete;
+        } catch (error) { throw error }
     }
 }
